create crud routes
