+++ server/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 // const { CrudOps } = require('../database-mongodb/itemAvailability.js');
-const { getData } = require('../database-pg/index.js');
+const { getData, deleteData, updateData, addData } = require('../database-couchdb/index.js');
 const cors = require('cors');
 const app = express();
 
@@ -17,22 +17,81 @@ app.get('*.js', (req, res, next) => {
 
 app.use(express.static(__dirname + '/../react-client/dist'));
 
-app.get('/availableAt/:itemId/', (req, res) => {
-  console.log('GET received! ID: ', req.params.itemId);
-  getData(req.params.itemId)
+app.get('/availableAt/:itemId', (req, res) => {
+  let itemId = req.params.itemId;
+  console.log('GET received! ID: ', itemId);
+
+  getData(itemId)
     .then((data) => {
-      // console.log('data: ', data);
+      console.log('data: ', data);
       if (Object.keys(data).length === 0) {
         res.sendStatus(404);
       } else {
-        res.status(200).send(data);
+        res.status(200).send(data.data.docs[0]);
       }
     })
     .catch((err) => {
       res.status(500).send(err);
     });
+
 });
 
+app.post('/availableAt', (req, res) => {
+  let itemId = req.params.itemId;
+  console.log('POST received! Data: ', req.body.data);
+
+  addData(itemId)
+    .then((data) => {
+      console.log('data: ', data);
+      if (Object.keys(data).length === 0) {
+        res.sendStatus(404);
+      } else {
+        res.status(200).send(data.data.docs[0]);
+      }
+    })
+    .catch((err) => {
+      res.status(500).send(err);
+    });
+
+});
+
+app.put('/availableAt/:itemId', (req, res) => {
+  let itemId = req.params.itemId;
+  console.log('Put received! ID: ', itemId);
+
+  updateData(itemId)
+    .then((data) => {
+      console.log('data: ', data);
+      if (Object.keys(data).length === 0) {
+        res.sendStatus(404);
+      } else {
+        res.status(200).send(data.data.docs[0]);
+      }
+    })
+    .catch((err) => {
+      res.status(500).send(err);
+    });
+
+});
+
+app.delete('/availableAt/:itemId', (req, res) => {
+  let itemId = req.params.itemId;
+  console.log('DELETE received! ID: ', itemId);
+
+  deleteData(itemId)
+    .then((data) => {
+      console.log('data: ', data);
+      if (Object.keys(data).length === 0) {
+        res.sendStatus(404);
+      } else {
+        res.status(200).send(data.data.docs[0]);
+      }
+    })
+    .catch((err) => {
+      res.status(500).send(err);
+    });
+
+});
 
 // //Create
 // app.post('/availableAt', (req, res) => {
